Un client Constellation.

Cette classe se connecte à un serveur Constellation déjà ouvert. Nous vous recommandons de ne pas l'utiliser directement, mais plutôt d'appeler `constellationR::avecClientEtServeur`, ou bien `constellationR::avecClient`, lesquels s'occuperont de la création et de la fermeture du client pour vous.

Public fields

port

Le numéro du port local sur lequel le serveur est ouvert, et auquel le client se connectera.

Active bindings

port

Le numéro du port local sur lequel le serveur est ouvert, et auquel le client se connectera.

Methods

Public methods

Client$new()

Client$action()

Client$suivre()

Client$rechercher()

Client$appeler()

Client$enregistrerÉcoute()

Client$fermer()

Client$clone()

Method new()

Usage

Client$new(port)

Method action()

Usage

Client$action(fonction, paramètres = NULL)

Method suivre()

Usage

Client$suivre(fonction, paramètres = NULL, nomArgFonction = "f")

Method rechercher()

Usage

Client$rechercher(fonction, paramètres, nomArgFonction = "f")

Method appeler()

Usage

Client$appeler(fonction, paramètres = NULL, nomArgFonction = "f")

Method enregistrerÉcoute()

Usage

Client$enregistrerÉcoute(id, résoudre, rejeter, f = NULL)

Method fermer()

Usage

Client$fermer()

Method clone()

The objects of this class are cloneable with this method.

Usage

Client$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.